Remove CRSN code
