+++ types/src/account_config/resources/mod.rs
@@ -5,12 +5,12 @@ pub mod chain_id;
 pub mod coin_info;
 pub mod coin_store;
 pub mod core_account;
-pub mod crsn;
+pub mod sequence_info;
 pub mod transaction_validation;
 
 pub use chain_id::*;
 pub use coin_info::*;
 pub use coin_store::*;
 pub use core_account::*;
-pub use crsn::*;
+pub use sequence_info::*;
 pub use transaction_validation::*;
